refactor: Reorganize parser folder (#18)
